feat(replication): Use 16 parallel replication processes
Instead of using a dynamic number of processes (2 × number of cores),
always use 16. It seems a good value for replicating over the network.

This solves the problem of replications being too long since CouchDB 2.1
(which has a new replication algorithm, with exponential backoff for
replication retry).

Examples of replication time for a large database:
- CouchDB 2.0:
  - 2 replication processes:  00:35:21
  - 16 replication processes: 00:12:19
  - 32 replication processes: failed with error 500
- CouchDB 2.1:
  - 2 replication processes:  02:50:31
  - 8 replication processes:  00:44:50
  - 16 replication processes: 00:24:22
  - 32 replication processes: failed with error 500