feat(restore-replicate): Add option to reuse database if already exists
This commit adds an option (`--reuse-db-if-exist`) to `coucharchive` so that
it can continue a restore or a replicate job even if the target CouchDB server
already contains databases with the same name.

**Use case #1**
> Restore from an archive to `A` but the server already
contains a database with the same name as a database in the archive
(in case of a previously failed `restore` for instance).

_Without `--reuse-db-if-exist`_
-> `coucharchive` would exit with this error message:
`The database could not be created, the file already exists.`

_With `--reuse-db-if-exist`_
-> `coucharchive` will not fail if the database already exists and will
continue replicating it.

**Use case #2**
> Restore from an archive to `A` but the server already
contains a database with the same name, different documents but the
same number of documents.

_Without `--reuse-db-if-exist`_
-> `coucharchive` would exit with this error message:
`The database could not be created, the file already exists.`

_With `--reuse-db-if-exist`_
-> `coucharchive` will not fail if the database already exists and will
continue replicating it.

**Use case #3**
> Restore from an archive to `A` but the server already
contains a database with the same name, different documents and
not same number of documents.

_Without `--reuse-db-if-exist`_
-> `coucharchive` would exit with this error message:
`[DB Name]: replicated database has X docs, source has X-1`

_With `--reuse-db-if-exist`_
-> same behavior as without the option so nothing is changed

The program has the same behavior with the `replicate` command.