fix(replicate): Manually handle CouchDB replication bug
This commit handles document creation when the document is not created
by replication because of the bug described below.

**Bug**

A bug exists in CouchDB when replicating a document that where
created, replicated, deleted, re-created (with the same ID)
and replicated again.

We expect CouchDB to create the document on the last replication as it
would normally do but it doesn't.

Another user already opened an issue on the CouchDB issue tracker.

See: https://github.com/apache/couchdb/issues/1418

I have tested it locally and the bug reproduction procedure is available
on the issue.

To confirm this fix works I am just using `coucharchive` on the last
replication.