fix(replication): Don't crash on handling bug 1418 if missing doc
This commit fixes the code that handles bug #1418, it a special case
where the source database is growing (a new document was created during
replication). In such cases, `bug_1418_create_missing_documents()`
should check whether the missing document ever existed or not. If not,
it is not bug #1418 (previously coucharchive crashed with `TypeError:
'NoneType' object is not subscriptable`). In this case, a new repliction
should be attempted.

The bug can be easily reproduced with:

[terminal 0] coucharchive load -i ~/backup-500-dbs.tar.gz
[terminal 1] i=0; while true; do ((i++)); curl -X PUT $COUCH/$DB/test-$i -d '{}'; sleep 0.1; done
[terminal 2] coucharchive create --from $COUCH -o /dev/null