feat(CouchDB): Limit number of replication worker processes
Following commit a8d764e768 let's also limit the number of worker
processes (see [1]).

Our tests showed that by using a lower value (1 instead of 4)
backups are more likely to succeed.

An explanation to this can be that `coucharchive` runs **many**
replication at the same time and is already managing by itself the
number of workers it needs to run smoothly.

[1]: https://docs.couchdb.org/en/latest/config/replicator.html#replicator/worker_processes