feat(script): add output help when archive cannot un-tar in `tmp`
During archive un-tar, if `No space left on device` is raised, propose user to
raise the size of his `/tmp` folder.
(I firmly believed, for a couple of hours, that I had a RAM issue).
On Fedora, a command to raise the size of the temporary folder until the next
reboot is:
```
sudo mount -o remount,size=20G,noatime /tmp
```