feat(replication): Use run_queue to avoid CouchDB overload
A few commits have try to avoid or recover from error that happen when
compute limitations are reached:
[`feat(replication): Handle recoverable problems`]
[`fix(replication): timeout`]
[`feat(replication): Handle recoverable problems`]
[`feat(replication): Use a dynamic replication speed`]
[`feat(CouchDB): Try to avoid requests timeout on replication`]
[`feat(CouchDB): Limit number of replication worker processes`]

But all of these commit allow `coucharchive` to run while machines themselves
(source or target) may be at a breaking point.
Depending what process inside CouchDB start to fail, CouchDB could recover or
it could go wild.

A solution I propose, to control the load on machines, source and target, is
to use [CouchDB `/_node/{node-name}/_system` API], with `_local` as `node-name`.
Especially, the `run_queue` info returned by the API seems a good representation
of compute limitations.
[Citing Adam Kocoloski]:
```
The “run_queue” is the number of Erlang processes in a runnable state that are
not currently executing on a scheduler. When that value is greater than zero it
means the node is hitting some compute limitations. Seeing a small positive
value from time to time is no problem.
```
The CLI option `--no-run-queue-control` can be set to true to restore the
previous behavior, if CouchDB errors are not a big deal in some use case.

A timeout parameter of 60 secondes is used when calling
[CouchDB `/_node/{node-name}/_system` API] in case this API keep the socket open
and do not answer. It happen to myself only once over hundred of tests, but it
can be critical to avoid overload.

[Citing Adam Kocoloski]: http://mail-archives.apache.org/mod_mbox/couchdb-user/201906.mbox/<90FEEE0B-B2E5-4B72-A196-4A35E26C467B%40apache.org>
[CouchDB `/_node/{node-name}/_system` API]: https://docs.couchdb.org/en/stable/api/server/common.html#node-node-name-system
[`feat(replication): Handle recoverable problems`]: https://github.com/adrienverge/coucharchive/commit/d1c0941955bf64a2e7b6714c07ea516dca50c5e7
[`fix(replication): timeout`]: https://github.com/adrienverge/coucharchive/commit/6c0863246f4da014bd2b93e6c0cd80b9e8f747cc
[`feat(replication): Handle recoverable problems`]: https://github.com/adrienverge/coucharchive/commit/4a326fe9bcccd326434e8cc98d8f0539f817bb5f
[`feat(replication): Use a dynamic replication speed`]: https://github.com/adrienverge/coucharchive/commit/1ea19faf72dd9442ecd1eb7a89e954b1c4c687a3
[`feat(CouchDB): Try to avoid requests timeout on replication`]: https://github.com/adrienverge/coucharchive/commit/a8d764e768a4529045c19b539ea2ad8ad8426555
[`feat(CouchDB): Limit number of replication worker processes`]: https://github.com/adrienverge/coucharchive/commit/b8125a929879d3180667bfb3804509c1971201ae