feat(replication): Implement a replication controller
The `ReplicationControl` object will be informed of past successes and
errors, and will dynamically adapt the number of concurrent replications
and timeout values.

The number of concurrent replications will start at 4, then will
continuously increase (with a maximum of 128), until an error is
detected. After that, it should stabilize around the last known value
without any error.