feat(replication): Implement --ideal-duration
This new parameter is available for `create`, `restore` and `replicate`;
it allows speeding down replication in order to decrease the overall
servers load.

For example on a server where replication would take 20 minutes, passing
`--ideal-duration 3600` will reduce the number of concurrent
replications, in order to finish in approximately one hour.