feat(replication): Average last error values, weighted by their age
Since commit [`feat(replication): Implement queues and replication controller`],
ideal number of concurrent replications is computed from last success and last
errors of replications.
When there have been errors in the last five minutes, it use the number of
concurrent replication at that time * 0.9.

It can create an oscillation effect:
- if multiple error happen
- ideal number of worker decrease progressively
- because the last value that had error is low, ideal number stay low for 5min
- passed that delay, the low value is forgotten so the ideal number of worker
  raise quickly
- multiple errors start to happen.

The explanation of it is that only the lowest last error is used, and this one
can be very low.
Also when a low value is forgotten a jump is created.
It creates a shuttering effect.

To avoid that effect, and stay just above the computationnal limit of the
machine, and still adapt to the size of dbs, let's forget progressively the
last errors, thanks to a age weighted average, and keep the
`_ideal_number_of_replications_for_ideal_duration` as a target.

[`feat(replication): Implement queues and replication controller`]: https://github.com/adrienverge/coucharchive/commit/45fbdd5

Tested with the following, with the source being a cluster of 3 vms and 70k dbs:
```
> coucharchive -v create --from toto -o toto.tar.gz --ideal-duration 10800
```