fix(replication): Performance drop following v2.2.0 part 2
Following commit [`feat(replication): Implement queues and replication
controller`], with a `MAX_NUMBER_OF_WORKERS = 64` configuration, performance
have dropped.
It is said in the commit that the script will take ~15% more time.
But, in production, with ~70k small dbs, we see a worse increase of ~26%.

Re-doing the test of [`feat(replication): Implement queues and replication
controller`], I see that first part of the drop is due to
`ideal_number_of_replications()` computation, and the second part is due
to delay introduce by the main loop.

This commit fix the second part by reducing the sleeping time of the main loop.
Result on a 3k small dbs cluster:
- v2.1.6:
  - 28,44s user 6,77s system 24% cpu 2:26,02 total
  - 28,13s user 6,68s system 23% cpu 2:28,11 total
- this PR*:
  - 21,98s user 3,83s system 17% cpu 2:25,52 total
  - 22,44s user 3,79s system 17% cpu 2:27,15 total
- previous commit (v2.2.1+first part of the fix)*:
  - 20,03s user 3,65s system 13% cpu 2:50,08 total
  - 20,02s user 3,63s system 14% cpu 2:48,63 total

*For this PR and previous commit tests the following patch is used so results
can be compared to v2.1.6:
```diff
--- a/coucharchive
+++ b/coucharchive
@@ -34,7 +34,7 @@ import urllib.request
 import couchdb

-MAX_NUMBER_OF_WORKERS = 128
+MAX_NUMBER_OF_WORKERS = 64

 def _canonical_couchdb_url(url):
@@ -296,7 +296,7 @@ class ReplicationControl(object):
         if not len(self._last_successes):
             # We do not have enough data yet to return anything useful.
             # Start with 4 concurrent replications:
-            return 4
+            return 64
         else:
             # Compute the ideal number of concurrent replications, to finish in
             # ideal_duration:
```

[`feat(replication): Implement queues and replication controller`]: https://github.com/adrienverge/coucharchive/commit/45fbdd5