feat: Raise max number of files (ulimit -n) to 65536
On most systems, the default maximum number of open file descriptors is
too low to work on a big archive. The archive loads, but trying to
operate heavily on its bases raises errors 500:

	(500, ('internal_server_error', 'No DB shards could be opened.'))

On my computer `ulimit -n` gives 1024, whereas the system maximum
(`ulimit -n unlimited`) is 524288. I propose to use 65536, which is a
convenient trade-off.